Change auth to store
